// src/@types/stock.d.ts
export default interface Stock {
  "52WeekChange": number;
  SandP52WeekChange: number;
  address1: string;
  algorithm?: any;
  annualHoldingsTurnover?: any;
  annualReportExpenseRatio?: any;
  ask: number;
  askSize: number;
  averageDailyVolume10Day: number;
  averageVolume: number;
  averageVolume10days: number;
  beta: number;
  beta3Year?: any;
  bid: number;
  bidSize: number;
  bookValue: number;
  category?: any;
  circulatingSupply?: any;
  city: string;
  companyOfficers: any[];
  country: string;
  currency: string;
  dateShortInterest: number;
  dayHigh: number;
  dayLow: number;
  dividendRate: number;
  dividendYield: number;
  earningsQuarterlyGrowth: number;
  enterpriseToEbitda: number;
  enterpriseToRevenue: number;
  enterpriseValue: number;
  exDividendDate: number;
  exchange: string;
  exchangeTimezoneName: string;
  exchangeTimezoneShortName: string;
  expireDate?: any;
  fiftyDayAverage: number;
  fiftyTwoWeekHigh: number;
  fiftyTwoWeekLow: number;
  fiveYearAverageReturn?: any;
  fiveYearAvgDividendYield: number;
  floatShares: number;
  forwardEps: number;
  forwardPE: number;
  fromCurrency?: any;
  fullTimeEmployees: number;
  fundFamily?: any;
  fundInceptionDate?: any;
  gmtOffSetMilliseconds: string;
  heldPercentInsiders: number;
  heldPercentInstitutions: number;
  industry: string;
  isEsgPopulated: boolean;
  lastCapGain?: any;
  lastDividendValue?: any;
  lastFiscalYearEnd: number;
  lastMarket?: any;
  lastSplitDate: number;
  lastSplitFactor: string;
  legalType?: any;
  logo_url: string;
  longBusinessSummary: string;
  longName: string;
  market: string;
  marketCap: number;
  maxAge: number;
  maxSupply?: any;
  messageBoardId: string;
  morningStarOverallRating?: any;
  morningStarRiskRating?: any;
  mostRecentQuarter: number;
  navPrice?: any;
  netIncomeToCommon: number;
  nextFiscalYearEnd: number;
  open: number;
  openInterest?: any;
  payoutRatio: number;
  pegRatio: number;
  phone: string;
  previousClose: number;
  priceHint: number;
  priceToBook: number;
  priceToSalesTrailing12Months: number;
  profitMargins: number;
  quoteType: string;
  regularMarketDayHigh: number;
  regularMarketDayLow: number;
  regularMarketOpen: number;
  regularMarketPreviousClose: number;
  regularMarketPrice: number;
  regularMarketVolume: number;
  revenueQuarterlyGrowth?: any;
  sector: string;
  sharesOutstanding: number;
  sharesPercentSharesOut: number;
  sharesShort: number;
  sharesShortPreviousMonthDate: number;
  sharesShortPriorMonth: number;
  shortName: string;
  shortPercentOfFloat: number;
  shortRatio: number;
  startDate?: any;
  state: string;
  strikePrice?: any;
  symbol: string;
  threeYearAverageReturn?: any;
  toCurrency?: any;
  totalAssets?: any;
  tradeable: boolean;
  trailingAnnualDividendRate: number;
  trailingAnnualDividendYield: number;
  trailingEps: number;
  trailingPE: number;
  twoHundredDayAverage: number;
  volume: number;
  volume24Hr?: any;
  volumeAllCurrencies?: any;
  website: string;
  yield?: any;
  ytdReturn?: any;
  zip: string;
}